move some stuff to defs, move TimeFromNowField to fields.py, fix nits
