+++ cabot/cabotapp/defs.py
@@ -27,3 +27,9 @@ DEFAULT_TCP_TIMEOUT = 8
 MAX_TCP_TIMEOUT = 16
 
 TIMESTAMP_FORMAT = "%b %d, %Y, %H:%M %Z"  # Eg: "Dec 10, 2018, 21:34 UTC"
+
+# options to present in the dropdown on the for creating an ack
+EXPIRE_AFTER_HOURS_OPTIONS = [1, 2, 4, 8, 24]
+
+# number of closed acks to show at the bottom of the acknowledgements page
+NUM_VISIBLE_CLOSED_ACKS = 12
